feat: relocated module app
