--- src/app/app.component.css
@@ -1,3 +1,0 @@
-:host() {
-    @apply w-screen h-screen flex items-center justify-center;
-}
