--- src/app/app.component.html
@@ -1,3 +1,0 @@
-<h1 class="text-red-400">
-  EN CONSTRUCCION
-</h1>
